Refactored out common delay logic and initialization
